169 - Rename buttons folder to testFilters and relocate buttons' styles file
