+++ frontend/src/pages/Overview.jsx
@@ -6,8 +6,8 @@ import Build from '../components/overview/Build';
 import Series from '../components/overview/Series';
 import ParentBuild from '../components/parentData/ParentBuild';
 import { ParentInfo } from './Overview.styles';
-import BuildAmountSelector from '../components/buttons/BuildAmountSelector';
-import Offset from '../components/buttons/OffSetButtons';
+import BuildAmountSelector from '../components/testFilters/BuildAmountSelector';
+import Offset from '../components/testFilters/OffSetButtons';
 import ContentHeader from '../components/header/ContentHeader';
 import { ContainerGrid12, ContentGrid6 } from '../styles/baseComponents';
 import ParentSeries from '../components/parentData/ParentSeries';
