+++ frontend/src/pages/SeriesHistory.jsx
@@ -2,9 +2,9 @@ import React, { Fragment, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import SeriesTestResultTable from '../components/seriesTestResultTable/SeriesTestResultTable';
 import ParentSeries from '../components/parentData/ParentSeries';
-import Offset from '../components/buttons/OffSetButtons';
-import LastRunCheckbox from '../components/buttons/LastRunCheckbox';
-import BuildAmountSelector from '../components/buttons/BuildAmountSelector';
+import Offset from '../components/testFilters/OffSetButtons';
+import LastRunCheckbox from '../components/testFilters/LastRunCheckbox';
+import BuildAmountSelector from '../components/testFilters/BuildAmountSelector';
 import { useStateValue } from '../contexts/state';
 import { useParams } from 'react-router';
 import BreadcrumbNav from '../components/BreadcrumbNav';
